Update GitLab CI
We now run CI for merge-request pipelines as well.
This has the advantage that it supports merge trains, for example.

But we need to make sure that we do not run CI twice for a single push
if a branch has an MR open.

Furthermore, GitLab advises against mixing rules and only/except,
so we switch all the only/except definitions for jobs to rules as well.